fix: grand total amount format in template generation
