+++ src/main/java/it/gov/pagopa/receipt/pdf/helpdesk/service/impl/BizEventToReceiptServiceImpl.java
@@ -27,12 +27,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static it.gov.pagopa.receipt.pdf.helpdesk.utils.BizEventToReceiptUtils.formatAmount;
+import static it.gov.pagopa.receipt.pdf.helpdesk.utils.BizEventToReceiptUtils.getAmount;
 import static it.gov.pagopa.receipt.pdf.helpdesk.utils.BizEventToReceiptUtils.getItemSubject;
 
 public class BizEventToReceiptServiceImpl implements BizEventToReceiptService {
@@ -258,17 +257,6 @@ public class BizEventToReceiptServiceImpl implements BizEventToReceiptService {
 
         receipt.setEventData(eventData);
         return receipt;
-    }
-
-    public static BigDecimal getAmount(BizEvent bizEvent) {
-        if (bizEvent.getTransactionDetails() != null && bizEvent.getTransactionDetails().getTransaction() != null
-                && bizEvent.getTransactionDetails().getTransaction().getGrandTotal() != 0) {
-            return formatAmount(bizEvent.getTransactionDetails().getTransaction().getGrandTotal());
-        }
-        if (bizEvent.getPaymentInfo() != null && bizEvent.getPaymentInfo().getAmount() != null) {
-            return new BigDecimal(bizEvent.getPaymentInfo().getAmount());
-        }
-        return BigDecimal.ZERO;
     }
 
 }
